LFS-654: Better UI for Questionnaire wizard Further refactoring to avoid code duplication
